change lat and long to floatfield

+++ main_app/models.py
@@ -1,4 +1,6 @@
 from django.db import models
+from django.contrib.auth.models import User
+
 # Create your models here.
 
 
@@ -6,8 +8,8 @@ class Residence(models.Model):
     address_line_1 = models.CharField(max_length=100)
     address_line_2 = models.CharField(max_length=100)
     city = models.CharField(max_length=50)
-    latitude = models.IntegerField()
-    longitude = models.IntegerField()
+    latitude = models.FloatField()
+    longitude = models.FloatField()
     start_date = models.DateField()
     end_date = models.DateField()
 
@@ -16,8 +18,8 @@ class Workplace(models.Model):
     address_line_1 = models.CharField(max_length=100)
     address_line_2 = models.CharField(max_length=100)
     city = models.CharField(max_length=50)
-    latitude = models.IntegerField()
-    longitude = models.IntegerField()
+    latitude = models.FloatField()
+    longitude = models.FloatField()
     company_name = models.CharField(max_length=100)
     start_date = models.DateField()
     end_date = models.DateField()
@@ -31,5 +33,5 @@ class Workplace(models.Model):
 #     address_line_1 = models.CharField(max_length=100)
 #     address_line_2 = models.CharField(max_length=100)
 #     city = models.CharField(max_length=50)
-#     latitude = models.IntegerField()
-#     longitude = models.IntegerField()
+#     latitude = models.FloatField()
+#     longitude = models.FloatField()
